docs(layout): re-review polish — cap approximations owned in comments
- cappedRatioWrite documents the raw-span approximation (safe
  direction: never blocks the visible bound) and the deliberate
  no-cap branch on too-narrow displays
- PerSpaceResizeTests notes the screen-width assumption the
  grow assertions now carry
- lua-reference stack resize bullet mentions the write stopping
  at the min_window_size bound
- split StackLayout domain authorities into StackLayout+Domain
  (main file crossed the 350-line ceiling)
