+++ Sources/KiwiDeskCore/Layouts/StackLayout.swift
@@ -8,123 +8,6 @@ import CoreGraphics
 public struct StackLayout: LayoutSystem {
     public init() {}
 
-    /// Renormalization floor applied when *reading* a stack
-    /// weight (#67) — shared with the `resize` command so the
-    /// command's step math and the layout's distribution can
-    /// never disagree on the domain. Deliberately below
-    /// `weightRange.lowerBound`: the command never stores a
-    /// value this small, so the floor only defends against a
-    /// future writer — do not collapse the two constants.
-    public static let weightFloor: Double = 0.05
-    /// Clamp for what the `resize` command *stores* (#67).
-    public static let weightRange: ClosedRange<Double> = 0.1...10
-
-    /// The master_ratio interval that keeps BOTH zones at
-    /// least `minSize` wide within `available` points (#44),
-    /// or nil when two min-size zones cannot coexist at any
-    /// ratio — the cascade case. `minSize` ≤ 0 (or NaN) means
-    /// no minimum: the full 0...1, which also fences nonsense
-    /// stored ratios away from negative zone widths. The
-    /// single authority behind the layout's effective clamp
-    /// and the interactive write cap (parity rule).
-    public static func effectiveRatioRange(
-        available: Double,
-        minSize: Double
-    ) -> ClosedRange<Double>? {
-        // A degenerate region (gaps wider than the usable
-        // area) cascades regardless of the minimum.
-        guard available > 0 else { return nil }
-        guard minSize > 0 else { return 0...1 }
-        guard available >= minSize * 2 else { return nil }
-        let fraction = minSize / available
-        return fraction...(1 - fraction)
-    }
-
-    /// Caps an interactive master_ratio write at the current
-    /// display's effective range (#44): past it the layout
-    /// clamps anyway and the stored value would only ratchet
-    /// invisibly — the same rule as the #67 weight cap. Never
-    /// pushes the value back across `base`, so an already
-    /// out-of-range value stays editable in the direction
-    /// that re-enters the range. (The config verb
-    /// `stack.set_master_ratio` deliberately stays wide: the
-    /// stored value is honored again on a wider display.)
-    public static func cappedRatioWrite(
-        _ proposed: Double,
-        base: Double,
-        available: Double,
-        minSize: Double
-    ) -> Double {
-        guard
-            let range = effectiveRatioRange(
-                available: available,
-                minSize: minSize
-            )
-        else { return proposed }
-        if proposed > base {
-            return min(proposed, max(range.upperBound, base))
-        }
-        if proposed < base {
-            return max(proposed, min(range.lowerBound, base))
-        }
-        return proposed
-    }
-
-    /// The largest weight total a column can carry before its
-    /// smallest share drops below `minSize` — the single
-    /// authority behind both the layout's cascade check and
-    /// the resize command's growth cap, so the two formulas
-    /// cannot drift apart (#67 review; parity rule).
-    public static func maxColumnTotal(
-        smallestWeight: Double,
-        height: Double,
-        minSize: Double
-    ) -> Double {
-        // No (or nonsense) minimum → no cliff, matching the
-        // old `share < minSize` comparison for minSize ≤ 0.
-        guard minSize > 0 else { return .infinity }
-        return smallestWeight * height / minSize
-    }
-
-    /// The master/stack partition of a tiled window array —
-    /// the single authority consumed by `calculateGeometry`
-    /// and the `resize` command (#67 review: a third
-    /// hand-mirror of this rule had crept in). `stack` is nil
-    /// while everything still fits in the master zone.
-    public static func partition(
-        _ windows: [WindowID],
-        masterCount: Int
-    ) -> (
-        master: ArraySlice<WindowID>,
-        stack: ArraySlice<WindowID>?
-    ) {
-        let boundary = max(1, masterCount)
-        guard windows.count > boundary else {
-            return (windows[...], nil)
-        }
-        return (windows[..<boundary], windows[boundary...])
-    }
-
-    /// The column (zone) of `partition` holding `member`, or
-    /// nil when it is not in `windows`.
-    public static func column(
-        containing member: WindowID,
-        in windows: [WindowID],
-        masterCount: Int
-    ) -> ArraySlice<WindowID>? {
-        guard let index = windows.firstIndex(of: member) else {
-            return nil
-        }
-        let (master, stack) = partition(
-            windows,
-            masterCount: masterCount
-        )
-        guard let stack, index >= master.endIndex else {
-            return master
-        }
-        return stack
-    }
-
     public func calculateGeometry(
         for windows: [WindowID],
         in context: LayoutContext
